requirements.txt
streamlit==1.24.0
psycopg2-binary==2.9.6
pandas==2.0.3
openai==1.3.0
python-dotenv==1.0.0
####